Extract SafariView to its own file
