+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/CardReadersView.swift
@@ -1,19 +1,5 @@
 import SwiftUI
-import SafariServices
 
-struct SafariView: UIViewControllerRepresentable {
-
-    let url: URL
-
-    func makeUIViewController(context: UIViewControllerRepresentableContext<SafariView>) -> SFSafariViewController {
-        return SFSafariViewController(url: url)
-    }
-
-    func updateUIViewController(_ uiViewController: SFSafariViewController,
-                                context: UIViewControllerRepresentableContext<SafariView>) {
-
-    }
-}
 /// A view to be displayed on Card Reader Manuals screen
 ///
 struct CardReadersView: View {
