+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/In-Person Payments/SafariView.swift
@@ -1,0 +1,19 @@
+import SwiftUI
+import SafariServices
+
+/// SwiftUI interface for UIKit SFSafariViewController
+/// Provides a visible interface for web browsing, and Safari features
+///
+struct SafariView: UIViewControllerRepresentable {
+
+    let url: URL
+
+    func makeUIViewController(context: UIViewControllerRepresentableContext<SafariView>) -> SFSafariViewController {
+        return SFSafariViewController(url: url)
+    }
+
+    func updateUIViewController(_ uiViewController: SFSafariViewController,
+                                context: UIViewControllerRepresentableContext<SafariView>) {
+
+    }
+}
